Julia sets functionality added, but not enabled.
The Julia sets I have found do not look as good as the
Mandelbrot/Burning Ship IMO. They appear to repeat too often to be
interesting, but maybe there is more interesting ones.

Additionally, I would need to implement the ability to set the Julia set
center in order for it to be usable, which I may do in the future.